--- a/host_core/lib/host_core/actors/actor_module.ex
+++ b/host_core/lib/host_core/actors/actor_module.ex
@@ -192,15 +192,16 @@ defmodule HostCore.Actors.ActorModule do
   end
 
   @impl true
-  def handle_cast(
+  def handle_call(
         {
           :handle_incoming_rpc,
           %{
             body: body,
-            reply_to: reply_to,
+            reply_to: _reply_to,
             topic: topic
           } = msg
         },
+        _from,
         agent
       ) do
     reconstitute_trace_context(Map.get(msg, :headers))
@@ -286,23 +287,11 @@ defmodule HostCore.Actors.ActorModule do
         end
 
       Task.start(fn ->
-        HostCore.Nats.safe_pub(
-          :lattice_nats,
-          reply_to,
-          ir |> Msgpax.pack!() |> IO.iodata_to_binary()
-        )
-      end)
-
-      # Tracer.add_event("Reply published", [])
-
-      Task.start(fn ->
         publish_invocation_result(inv, ir)
       end)
+
+      {:reply, {:ok, ir |> Msgpax.pack!() |> IO.iodata_to_binary()}, agent}
     end
-
-    # span
-
-    {:noreply, agent}
   end
 
   defp reconstitute_trace_context(headers) when is_list(headers) do