--- a/host_core/lib/host_core/actors/actor_rpc_server.ex
+++ b/host_core/lib/host_core/actors/actor_rpc_server.ex
@@ -18,10 +18,10 @@ defmodule HostCore.Actors.ActorRpcServer do
 
       actors ->
         next_index = CallCounter.read_and_increment(pk)
+        IO.puts(next_index)
         {pid, _value} = Enum.at(actors, rem(next_index, length(actors)))
 
         GenServer.cast(pid, {:handle_incoming_rpc, msg})
-        :ok
     end
   end