--- a/host_core/mix.exs
+++ b/host_core/mix.exs
@@ -43,8 +43,7 @@ defmodule HostCore.MixProject do
       {:rustler, "~> 0.24.0"},
       {:timex, "~> 3.7"},
       {:jason, "~> 1.2.2"},
-      # {:gnat, "~> 1.5.2"},
-      {:gnat, path: "../../../brooksmtownsend/nats.ex"},
+      {:gnat, "~> 1.5.2"},
       # erlavro isn't used, but this version upgrades dependency of cloudevents 0.4.0 to use rebar3
       {:erlavro, "~> 2.9.7", override: true, manager: :rebar3},
       {:cloudevents, "~> 0.4.0"},