… host_core/mix.lock
… host_core/test/fixtures/actors/echo.wasm